// src/apis/models/IczpNet_Chat_SessionSections_SessionUnits_SessionUnitCacheItem.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type IczpNet_Chat_SessionSections_SessionUnits_SessionUnitCacheItem = {
    id?: string;
    sessionId?: string | null;
    appUserId?: string | null;
    ownerId?: number;
    destinationId?: number | null;
    isPublic?: boolean;
    readedMessageId?: number | null;
    lastMessageId?: number | null;
    publicBadge?: number;
    privateBadge?: number;
    remindAllCount?: number;
    remindMeCount?: number;
    followingCount?: number;
    ticks?: number;
};